Server side first method not passing

+++ java/tenmo-server/src/main/java/com/techelevator/tenmo/controller/TenmoController.java
@@ -26,7 +26,7 @@ public class TenmoController {
 
 @RequestMapping(path = "/balance", method = RequestMethod.GET)
     public Account getAccount(Principal principal){
-    int id = userDao.findIdByUsername(principal.getName());
+    long id = userDao.findIdByUsername(principal.getName());
     return accountDao.getAccount(id);
   }
